fix(jobboard): Cast userid to int in audit log function
Moodle sometimes stores user IDs as strings. The audit::log() function
now explicitly casts the userid to int before passing it to
get_user_jobboard_role() which expects an integer parameter.

+++ local/jobboard/classes/audit.php
@@ -100,7 +100,8 @@ class audit {
     ) {
         global $DB, $USER;
 
-        $actualuserid = $userid ?? ($USER->id ?? 0);
+        // Ensure userid is always an integer (Moodle sometimes stores IDs as strings).
+        $actualuserid = (int) ($userid ?? ($USER->id ?? 0));
 
         $record = new \stdClass();
         $record->userid = $actualuserid;
